+++ src/blob/core.rs
@@ -1,12 +1,12 @@
 use tokio::time::Instant;
 
+use crate::error::ValidationErrorKind;
+
 use super::prelude::*;
 
-use super::index::IndexTrait;
+use super::{header::Header, index::IndexTrait};
 
-const BLOB_VERSION: u32 = 1;
-const BLOB_MAGIC_BYTE: u64 = 0xdeaf_abcd;
-const BLOB_INDEX_FILE_EXTENSION: &str = "index";
+pub(crate) const BLOB_INDEX_FILE_EXTENSION: &str = "index";
 
 /// A [`Blob`] struct representing file with records,
 /// provides methods for read/write access by key
@@ -71,8 +71,7 @@ impl Blob {
         if self.index.on_disk() {
             Ok(0) // 0 bytes dumped
         } else {
-            self.file
-                .fsyncdata()
+            self.fsyncdata()
                 .await
                 .with_context(|| "Blob file dump failed!")?;
             self.index
@@ -106,14 +105,32 @@ impl Blob {
         let name = FileName::from_path(&path)?;
         info!("{} blob init started", name);
         let size = file.size();
-        let header = Header::new();
-        Self::check_blob_header(&header)?;
+
+        let header = Header::from_file(&name, ioring.clone())
+            .await
+            .context("failed to read blob header")?;
+
         let mut index_name = name.clone();
         index_name.extension = BLOB_INDEX_FILE_EXTENSION.to_owned();
         trace!("looking for index file: [{}]", index_name);
+        let mut is_index_corrupted = false;
         let index = if index_name.exists() {
             trace!("file exists");
-            Index::from_file(index_name, index_config, ioring).await?
+            Index::from_file(index_name.clone(), index_config.clone(), ioring.clone())
+                .await
+                .or_else(|error| {
+                    if let Some(io_error) = error.downcast_ref::<IOError>() {
+                        match io_error.kind() {
+                            IOErrorKind::PermissionDenied | IOErrorKind::Other => {
+                                warn!("index cannot be regenerated due to error: {}", io_error);
+                                return Err(error);
+                            }
+                            _ => {}
+                        }
+                    }
+                    is_index_corrupted = true;
+                    Ok(Index::new(index_name, ioring, index_config))
+                })?
         } else {
             trace!("file not found, create new");
             Index::new(index_name, ioring, index_config)
@@ -128,7 +145,7 @@ impl Blob {
             current_offset: Arc::new(Mutex::new(size)),
         };
         trace!("call update index");
-        if size as u64 > header_size {
+        if is_index_corrupted || size as u64 > header_size {
             blob.try_regenerate_index(blob_key_size as usize)
                 .await
                 .context("failed to regenerate index")?;
@@ -183,14 +200,6 @@ impl Blob {
 
     pub(crate) const fn check_data_consistency() {
         // @TODO implement
-    }
-
-    fn check_blob_header(header: &Header) -> Result<()> {
-        if header.magic_byte == BLOB_MAGIC_BYTE {
-            Ok(())
-        } else {
-            Err(Error::validation("blob header magic byte is wrong").into())
-        }
     }
 
     pub(crate) async fn write(&mut self, mut record: Record) -> Result<()> {
@@ -405,29 +414,6 @@ impl Display for FileName {
     }
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct Header {
-    magic_byte: u64,
-    version: u32,
-    flags: u64,
-}
-
-impl Header {
-    pub const fn new() -> Self {
-        Self {
-            magic_byte: BLOB_MAGIC_BYTE,
-            version: BLOB_VERSION,
-            flags: 0,
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct Location {
-    offset: u64,
-    size: usize,
-}
-
 struct RawRecords {
     current_offset: u64,
     record_header_size: u64,
@@ -457,8 +443,8 @@ impl RawRecords {
         let key_len = bincode::deserialize::<usize>(key_len_buf)
             .context("failed to deserialize index buf vec length")?;
         if key_len != key_size {
-            let msg = "blob key_sizeis not equal to pearl compile-time key size";
-            return Err(Error::validation(msg).into());
+            let msg = "blob key_size is not equal to pearl compile-time key size";
+            return Err(Error::validation(ValidationErrorKind::BlobKeySize, msg).into());
         }
         let record_header_size = RecordHeader::default().serialized_size() + key_len as u64;
         debug!(
@@ -476,7 +462,8 @@ impl RawRecords {
         if magic_byte == RECORD_MAGIC_BYTE {
             Ok(())
         } else {
-            Err(Error::validation("First record's magic byte is wrong").into())
+            let param = ValidationErrorKind::RecordMagicByte;
+            Err(Error::validation(param, "First record's magic byte is wrong").into())
         }
     }