+++ src/record.rs
@@ -1,4 +1,4 @@
-use crate::prelude::*;
+use crate::{error::ValidationErrorKind, prelude::*};
 
 pub(crate) const RECORD_MAGIC_BYTE: u64 = 0xacdc_bcde;
 
@@ -118,7 +118,8 @@ impl Record {
         if self.header().magic_byte == RECORD_MAGIC_BYTE {
             Ok(())
         } else {
-            Err(Error::validation("wrong magic byte").into())
+            let param = ValidationErrorKind::RecordMagicByte;
+            Err(Error::validation(param, "wrong magic byte").into())
         }
     }
 
@@ -131,7 +132,8 @@ impl Record {
                 "wrong data checksum {} vs {}",
                 calc_crc, self.header.data_checksum
             );
-            let e = Error::validation(cause);
+            let param = ValidationErrorKind::RecordDataChecksum;
+            let e = Error::validation(param, cause);
             error!("{:#?}", e);
             Err(e.into())
         }
@@ -146,10 +148,12 @@ impl Record {
         if calc_crc == self.header.header_checksum {
             Ok(())
         } else {
-            let e = ErrorKind::Validation(format!(
+            let cause = format!(
                 "wrong header checksum {} vs {}",
                 calc_crc, self.header.header_checksum
-            ));
+            );
+            let param = ValidationErrorKind::RecordHeaderChecksum;
+            let e = Error::validation(param, cause);
             error!("{:#?}", e);
             Err(Error::from(e).into())
         }