+++ CHANGELOG.md
@@ -1,16 +1,38 @@
 # Changelog
 Pearl changelog
 
+
 ## [Unreleased]
 #### Added
-- B+ Tree indexes ([#84](https://github.com/qoollo/pearl/pull/93))
-- Range indexes ([#11](https://github.com/qoollo/pearl/pull/102))
+
 
 #### Changed
 
+
 #### Fixed
 
+
 #### Updated
+
+
+## [0.7.1] - 2021-10-18
+#### Added
+- Add more functions to interface to support work with optional active blob (#118)
+
+
+## [0.7.0] - 2021-10-02
+#### Added
+- B+ Tree indexes (#84)
+- Range indexes (#11)
+- Move corrupted blobs into speacial directory (#98)
+- Blob version validation (#120)
+
+#### Changed
+- Rebuild corrupted index automatically during startup (#94)
+- Move build_time.rs file to OUT_DIR.
+
+#### Fixed
+- Create corrupted directory only when needed (#94)
 
 
 ## [0.6.2] - 2021-09-14